# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: CoffeeBeansStock API
  description: コーヒー豆在庫管理アプリケーション API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      summary: トークンリフレッシュ
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: リフレッシュ成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResult"
        "401":
          description: トークン無効
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    RefreshResult:
      type: object
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
        refresh_token:
          type: string

    # ── Coffee Beans ─────────────────────────────────────
...