# backend/internal/api/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for RoastLevel.
const (
	Cinnamon RoastLevel = "cinnamon"
	City     RoastLevel = "city"
	French   RoastLevel = "french"
	FullCity RoastLevel = "full_city"
	High     RoastLevel = "high"
	Italian  RoastLevel = "italian"
	Light    RoastLevel = "light"
	Medium   RoastLevel = "medium"
)

// Valid indicates whether the value is a known member of the RoastLevel enum.
func (e RoastLevel) Valid() bool {
	switch e {
	case Cinnamon:
		return true
	case City:
		return true
	case French:
		return true
	case FullCity:
		return true
	case High:
		return true
	case Italian:
		return true
	case Light:
		return true
	case Medium:
		return true
	default:
		return false
	}
}

// AuthResponse defines model for AuthResponse.
type AuthResponse struct {
	AccessToken  string       `json:"access_token"`
	RefreshToken string       `json:"refresh_token"`
	User         UserResponse `json:"user"`
}

// CoffeeBeanResponse defines model for CoffeeBeanResponse.
type CoffeeBeanResponse struct {
	CreatedAt    time.Time          `json:"created_at"`
	CurrentStock int32              `json:"current_stock"`
	Id           openapi_types.UUID `json:"id"`
	Name         string             `json:"name"`
	Notes        *string            `json:"notes,omitempty"`
	Origin       *string            `json:"origin,omitempty"`
	RoastLevel   RoastLevel         `json:"roast_level"`
	UpdatedAt    time.Time          `json:"updated_at"`
}

// CreateBeanRequest defines model for CreateBeanRequest.
type CreateBeanRequest struct {
	CurrentStock int32      `json:"current_stock"`
	Name         string     `json:"name"`
	Notes        *string    `json:"notes,omitempty"`
	Origin       *string    `json:"origin,omitempty"`
	RoastLevel   RoastLevel `json:"roast_level"`
}

// DeleteResponse defines model for DeleteResponse.
type DeleteResponse struct {
	Message string `json:"message"`
}

// ErrorBody defines model for ErrorBody.
type ErrorBody struct {
	Code    string        `json:"code"`
	Details *[]FieldError `json:"details,omitempty"`
	Message string        `json:"message"`
}

// FieldError defines model for FieldError.
type FieldError struct {
	Field   string `json:"field"`
	Message string `json:"message"`
}

// ListBeansResponse defines model for ListBeansResponse.
type ListBeansResponse struct {
	Beans      []CoffeeBeanResponse `json:"beans"`
	Pagination PaginationResponse   `json:"pagination"`
}

// PaginationResponse defines model for PaginationResponse.
type PaginationResponse struct {
	HasMore bool  `json:"has_more"`
	Limit   int32 `json:"limit"`
	Offset  int32 `json:"offset"`
	Total   int64 `json:"total"`
}

// RefreshRequest defines model for RefreshRequest.
type RefreshRequest struct {
	RefreshToken string `json:"refresh_token"`
}

// RefreshResult defines model for RefreshResult.
type RefreshResult struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}

// RoastLevel defines model for RoastLevel.
type RoastLevel string

// UpdateBeanRequest defines model for UpdateBeanRequest.
type UpdateBeanRequest struct {
	CurrentStock *int32      `json:"current_stock,omitempty"`
	Name         *string     `json:"name,omitempty"`
	Notes        *string     `json:"notes,omitempty"`
	Origin       *string     `json:"origin,omitempty"`
	RoastLevel   *RoastLevel `json:"roast_level,omitempty"`
}

// UserResponse defines model for UserResponse.
type UserResponse struct {
	CreatedAt           time.Time          `json:"created_at"`
	Id                  openapi_types.UUID `json:"id"`
	LowStockThreshold   *int32             `json:"low_stock_threshold,omitempty"`
	Name                *string            `json:"name,omitempty"`
	NotificationEnabled *bool              `json:"notification_enabled,omitempty"`
	UpdatedAt           *time.Time         `json:"updated_at,omitempty"`
}

// ListBeansParams defines parameters for ListBeans.
type ListBeansParams struct {
	Limit  *int32 `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int32 `form:"offset,omitempty" json:"offset,omitempty"`
}

// RefreshTokenJSONRequestBody defines body for RefreshToken for application/json ContentType.
type RefreshTokenJSONRequestBody = RefreshRequest

// CreateBeanJSONRequestBody defines body for CreateBean for application/json ContentType.
type CreateBeanJSONRequestBody = CreateBeanRequest

// UpdateBeanJSONRequestBody defines body for UpdateBean for application/json ContentType.
type UpdateBeanJSONRequestBody = UpdateBeanRequest